pyproject.toml
[tool.poetry]
name = "schedule-to-cal"
version = "0.1.0"
description = ""
authors = ["Bhavya Muni <bhavyamuni@outlook.com>"]
readme = "README.md"
packages = [{include = "schedule_to_cal"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
asyncio = "^3.4.3"
google-api-python-client = "^2.70.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.8.0"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8  = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
